Fix bug where the preview scene object was somehow getting changed
No quite sure what was happening exactly. Some of the initialisation
arguments of the scene were changes by thing happening during the game.
Which caused problems when going back to the editor after exiting the
preview. Fixed it by using the JSON string version of the scene, which
is guaranteed to remain unchanged.